BACKEND/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
openpyxl==3.1.2
pytz==2023.4
sqlparse==0.4.4
